// src/generated/accounts/orderBook.ts
export type OrderBook = {
  discriminator: ReadonlyUint8Array;
  version: number;
  orderBookType: OrderBookType;
  apy: APY;
  loanTerms: BookLoanTerms;
  feePermillicentage: number;
  feeAuthority: Address;
};